+++ sources/MarkTrigger.m
@@ -34,7 +34,7 @@ typedef enum {
     return YES;
 }
 
-- (int)indexOfTag:(int)theTag {
+- (NSInteger)indexOfTag:(NSInteger)theTag {
     int i = 0;
     for (NSNumber *n in [self objectsSortedByValueInDict:[self menuItemsForPoupupButton]]) {
         if ([n intValue] == theTag) {
@@ -45,7 +45,7 @@ typedef enum {
     return -1;
 }
 
-- (int)tagAtIndex:(int)index {
+- (NSInteger)tagAtIndex:(NSInteger)index {
     int i = 0;
 
     for (NSNumber *n in [self objectsSortedByValueInDict:[self menuItemsForPoupupButton]]) {
